requirements.txt
fastapi
uvicorn
torch==2.3.1
torchvision==0.18.1
mediapipe==0.10.14
Pillow
numpy